Test IntoIterator<Item=RistrettoPoint> in verify
This commit is for testing the effect of `.cloned()` on verify
performance.

If the `RangeProof` struct held `CompressedRistretto` points,
we could try to decompress the points on the fly as we pass them into
the verification API.  This means that we would be able to pass directly
from the serialized `RangeProof` structs to filling a buffer with
lookup tables, with no extra allocations.

This would be especially helpful for large batch verifications.

However, we can't chain iterators with `Item=RistrettoPoint` and
`Item=&RistrettoPoint`.  So this means we have to do `.cloned()` on all
of the static points.  Since the clone is *per-item*, not a buffer-wise
clone, I think it's possible in theory for this copy to be optimized
out.

This commit is to provide a stable reference for testing whether this
has a bad performance impact.